move canvas buttons to separate file
